refactor: 기능 단위로 파일 위치 분리
- 직접적인 게임에 관련된 로직은 game
- 지엽적인 부분에만 관여하는 함수는 utils
