+++ src/game/displayResult.js
